// src/app/models/initial-catalogs.model.ts
import { InwshieldRow } from './in-wshield-row.model';
import { JobType } from './job-type.mode';
import { CostIntegration } from './cost-integration.model';
import { JobMaterialType } from './job-material-type.mode';
import { ProspectingType } from './prospecting-type.model';
import { RoleResource } from './role-resource.model';
import { User } from './user.model';
import { WarrantyType } from './warranty-type.model';
import { WarrantyOption } from './warranty-option.model';
import { ShingleTypesRemove } from './shingle_types_remove.model';
import { MeasuresMaterialType } from './measures-material-types.model';
import { WallMaterial } from './wall-material.model';
import { General } from './general.model';
import { Trademark } from './trademark.model';
import { MaterialColor } from './material-color.model';
import { Group } from './group.model';
import { GroupColor } from './group-color.model';
import { CostType } from './cost_type.model';
import { Compability } from './compability.model';
import { Expens } from './expens.model';
import { Upgrade } from './upgrade.model';
import { MaterialPrice } from './material-price.model';
import { ConceptType } from './concept_type.mdoel';
import { LaborPitchPrice } from './labor-pitch-price.model';
import { MaterialUnit } from './material-unit.model';
import { Suppiler } from './supplier.model';
import { InspectorCostType } from './inspector-cost-type.model';
import { LaborPrice } from './labor-price.model';
import { PriceList } from './price-list.model';
import { Warranty } from './warranty.model';

export interface InitialCatalogs {
  jobTypes: JobType[];
  jobMaterialTypes: JobMaterialType[];
  prospectingTypes: ProspectingType[];
  warrantyTypes: WarrantyType[];
  users: User[];
  roleResources: RoleResource[];
  inwshieldRows: InwshieldRow[];
  costIntegrations: CostIntegration[];
  warrantyOptions: WarrantyOption[];
  shingleTypesRemoves: ShingleTypesRemove[];
  materialTypes: MeasuresMaterialType[];
  wallMaterial: WallMaterial[];
  general: General[];
  trademarks: Trademark[];
  materialColors: MaterialColor[];
  groups: Group[];
  groupColors: GroupColor[];
  costTypes: CostType[];
  compatibilities: Compability[];
  expenses: Expens[];
  upgrades: Upgrade[];
  materialPrices: MaterialPrice[];
  conceptTypes: ConceptType[];
  laborPrices: LaborPrice[];
  materialUnits: MaterialUnit[];
  suppilers: Suppiler[];
  inspectorCostTypes: InspectorCostType[];
  laborPitchPrices: LaborPitchPrice[];
  priceList: PriceList[];
  warranties: Warranty[];
}